+++ packages/app/babel.config.js
@@ -4,7 +4,6 @@ module.exports = function (api) {
     return {
         presets: [['babel-preset-expo', { unstable_transformImportMeta: true }]],
         plugins: [
-            'macros',
             '@babel/plugin-transform-class-static-block',
             'react-native-worklets/plugin',
             '@lingui/babel-plugin-lingui-macro',